Rename window::window_style to window::style
